+++ app/Http/Controllers/ProjectController.php
@@ -21,7 +21,10 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        $projects = Project::orderBy('created_at' ,'DESC')->get();
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
+        $projects = Project::orderBy('created_at', 'DESC')->get();
         return view('backend.projects.index', compact('projects'));
     }
 
@@ -32,6 +35,9 @@ class ProjectController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
         return view('backend.projects.create');
     }
 
@@ -43,35 +49,36 @@ class ProjectController extends Controller
      */
     public function store(StoreProjectRequest $request)
     {
-               // dd($request->all());
-               if($request->hasFile('image')){
-                // Get filename with the extension
-                $filenameWithExt = $request->file('image')->getClientOriginalName();
-                // Get just filename
-                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
-                // Get just ext
-                $extension = $request->file('image')->getClientOriginalExtension();
-                // Filename to store
-                $fileNameToStore= $filename.'_'.time().'.'.$extension;
-                // Upload Image
-    
-                $path = $request->file('image')->move(public_path('/uploads/projects'), $fileNameToStore);
-    
-            }
-            $project = new project();
-    
-             $project->title = $request->input('title');
-             $project->body = $request->input('body');
-             $project->slug = make_slug($request->input('title'));
-             $project->user_id = auth()->user()->id;
-             
-             $project->image = $fileNameToStore;
-             $project->save();
-    
-            Session::flash('success', 'Added successfully');
-        
-            return redirect(route('admin.projects.index'));
-            
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
+        // dd($request->all());
+        if ($request->hasFile('image')) {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload Image
+
+            $path = $request->file('image')->move(public_path('/uploads/projects'), $fileNameToStore);
+        }
+        $project = new project();
+
+        $project->title = $request->input('title');
+        $project->body = $request->input('body');
+        $project->slug = make_slug($request->input('title'));
+        $project->user_id = auth()->user()->id;
+
+        $project->image = $fileNameToStore;
+        $project->save();
+
+        Session::flash('success', 'Added successfully');
+
+        return redirect(route('admin.projects.index'));
     }
 
     /**
@@ -82,6 +89,9 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
         return view('backend.projects.view', compact('project'));
     }
 
@@ -93,7 +103,10 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
-        return view('backend.projects.create',compact('project'));
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
+        return view('backend.projects.create', compact('project'));
     }
 
     /**
@@ -105,49 +118,50 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
-            // dd($request->all());
-           //TODO: check validation and status button
-             
-             if($request->hasFile('image')){
-                // Get filename with the extension
-                $filenameWithExt = $request->file('image')->getClientOriginalName();
-                // Get just filename
-                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
-                // Get just ext
-                $extension = $request->file('image')->getClientOriginalExtension();
-                // Filename to store
-                $fileNameToStore= $filename.'_'.time().'.'.$extension;
-                // Upload Image
-    
-                $path = $request->file('image')->move(public_path('/uploads/projects'), $fileNameToStore);
-            }
- 
-         
-         $project->body = $request->input('body');
-         $project->user_id = auth()->user()->id;
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
+        // dd($request->all());
+        //TODO: check validation and status button
 
-         if($request->title) {
+        if ($request->hasFile('image')) {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload Image
+
+            $path = $request->file('image')->move(public_path('/uploads/projects'), $fileNameToStore);
+        }
+
+
+        $project->body = $request->input('body');
+        $project->user_id = auth()->user()->id;
+
+        if ($request->title) {
             $project->title = $request->input('title');
-             $project->slug = make_slug($request->input('title'));
-           }
+            $project->slug = make_slug($request->input('title'));
+        }
 
-           if($request->hasFile('image')){
+        if ($request->hasFile('image')) {
 
 
             $path = parse_url($project->image);
-    
+
             File::delete(public_path($path['path']));
-    
-                $project->image = $fileNameToStore;
-            }
-           $project->save();
-   
-         Session::flash('success', 'Updated Successfully');
- 
-         // redirect user
-         return redirect(route('admin.projects.index'));
- 
- 
+
+            $project->image = $fileNameToStore;
+        }
+        $project->save();
+
+        Session::flash('success', 'Updated Successfully');
+
+        // redirect user
+        return redirect(route('admin.projects.index'));
     }
 
 
@@ -159,13 +173,15 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        try{
+        if (!Gate::allows('is_admin')) {
+            abort(403);
+        }
+        try {
             $project->delete();
-            
+
             session()->flash('success', 'Deleted successfully');
-    
+
             return redirect(route('admin.projects.index'));
-            
         } catch (Exception $e) {
 
             return back()->withError($e->getMessage());
